+++ app/src/main/java/woowacourse/shopping/data/datasource/MypageDataSource.kt
@@ -1,0 +1,8 @@
+package woowacourse.shopping.data.datasource
+
+import woowacourse.shopping.data.remote.result.DataResult
+
+interface MypageDataSource {
+    fun getCash(callback: (DataResult<Int>) -> Unit)
+    fun chargeCash(cash: Int, callback: (DataResult<Int>) -> Unit)
+}
