+++ app/src/main/java/woowacourse/shopping/data/datasource/RecentViewedDataSource.kt
@@ -1,0 +1,9 @@
+package woowacourse.shopping.data.datasource
+
+import woowacourse.shopping.domain.model.Product
+
+interface RecentViewedDataSource {
+    fun findAll(callback: (List<Product>) -> Unit)
+    fun add(product: Product)
+    fun remove(id: Int)
+}
