+++ app/src/main/java/woowacourse/shopping/data/datasource/ServerStoreDataSource.kt
@@ -1,0 +1,6 @@
+package woowacourse.shopping.data.datasource
+
+interface ServerStoreDataSource {
+    fun setServerUrl(url: String)
+    fun getServerUrl(): String
+}
