+++ app/src/main/java/woowacourse/shopping/data/remote/OrderApi.kt
@@ -6,7 +6,6 @@ import retrofit2.http.GET
 import retrofit2.http.POST
 import retrofit2.http.Path
 import woowacourse.shopping.data.remote.dto.OrderCartItemsDTO
-import woowacourse.shopping.data.remote.dto.OrderSubmitDTO
 import woowacourse.shopping.data.remote.dto.OrdersDTO
 
 interface OrderApi {
@@ -14,7 +13,7 @@ interface OrderApi {
     fun requestOrders(): Call<OrdersDTO>
 
     @GET("/orders/{id}")
-    fun requestOrderDetail(@Path("id") orderId: Int): Call<OrderSubmitDTO>
+    fun requestOrderDetail(@Path("id") orderId: Int): Call<OrdersDTO.OrderDTO>
 
     @POST("/orders")
     fun requestOrderCartItems(@Body orderCartItemDtos: OrderCartItemsDTO): Call<Unit>
