+++ app/src/main/java/woowacourse/shopping/data/remote/dto/OrdersDTO.kt
@@ -1,13 +1,13 @@
 package woowacourse.shopping.data.remote.dto
 
-data class OrdersDTO(val orders: List<OrderDTO?>?) {
+data class OrdersDTO(val orders: List<OrderDTO>?) {
     val isNotNull: Boolean
         get() = orders != null && orders.all { it?.isNotNull ?: false }
 
     data class OrderDTO(
         val orderId: Int?,
         val orderedDateTime: String?,
-        val products: List<ProductWithQuantityDTO?>?,
+        val products: List<ProductWithQuantityDTO>?,
         val totalPrice: Int?,
     ) {
         val isNotNull: Boolean
