+++ app/src/main/java/woowacourse/shopping/data/remote/dto/ProductWithQuantityDTO.kt
@@ -1,6 +1,6 @@
 package woowacourse.shopping.data.remote.dto
 
-data class ProductWithQuantityDTO(val productDTO: ProductDTO?, val quantity: Int?) {
+data class ProductWithQuantityDTO(val product: ProductDTO?, val quantity: Int?) {
     val isNotNull: Boolean
-        get() = productDTO != null && productDTO.isNotNull && quantity != null
+        get() = product != null && product.isNotNull && quantity != null
 }
