+++ app/src/main/java/woowacourse/shopping/data/repository/impl/RecentViewedDbRepository.kt
@@ -1,33 +1,21 @@
 package woowacourse.shopping.data.repository.impl
 
-import android.content.Context
-import woowacourse.shopping.data.db.RecentViewedDBHelper
+import woowacourse.shopping.data.datasource.RecentViewedDataSource
 import woowacourse.shopping.data.repository.RecentViewedRepository
-import woowacourse.shopping.data.repository.ServerStoreRespository
 import woowacourse.shopping.domain.model.Product
 
 class RecentViewedDbRepository(
-    context: Context,
-    serverRepository: ServerStoreRespository,
+    private val recentViewedDataSource: RecentViewedDataSource,
 ) : RecentViewedRepository {
-    private val dbHelper = RecentViewedDBHelper(context, serverRepository)
-
     override fun findAll(callback: (List<Product>) -> Unit) {
-        callback(dbHelper.selectAll())
+        recentViewedDataSource.findAll(callback)
     }
 
     override fun add(product: Product) {
-        if (find(product.id) != null) {
-            dbHelper.remove(product.id)
-        }
-        dbHelper.insert(product)
-    }
-
-    private fun find(id: Int): Product? {
-        return dbHelper.selectWhereId(id)
+        recentViewedDataSource.add(product)
     }
 
     override fun remove(id: Int) {
-        return dbHelper.remove(id)
+        recentViewedDataSource.remove(id)
     }
 }
