+++ app/src/main/java/woowacourse/shopping/data/repository/impl/ServerPreferencesRepository.kt
@@ -1,20 +1,13 @@
 package woowacourse.shopping.data.repository.impl
 
-import android.content.Context
-import androidx.preference.PreferenceManager
+import woowacourse.shopping.data.datasource.ServerStoreDataSource
 import woowacourse.shopping.data.repository.ServerStoreRespository
 
-class ServerPreferencesRepository(context: Context) : ServerStoreRespository {
-    private val sharedPreferences = PreferenceManager.getDefaultSharedPreferences(context)
-
+class ServerPreferencesRepository(private val serverStoreDataSource: ServerStoreDataSource) :
+    ServerStoreRespository {
     override fun setServerUrl(url: String) {
-        val editor = sharedPreferences.edit()
-        editor.putString(SERVER, url).apply()
+        serverStoreDataSource.setServerUrl(url)
     }
 
-    override fun getServerUrl(): String = sharedPreferences.getString(SERVER, "") ?: ""
-
-    companion object {
-        const val SERVER = "SERVER"
-    }
+    override fun getServerUrl(): String = serverStoreDataSource.getServerUrl()
 }
