+++ app/src/test/java/woowacourse/shopping/presenter/OrderCartProductModelFixture.kt
@@ -1,4 +1,4 @@
-package woowacourse.shopping
+package woowacourse.shopping.presenter
 
 import woowacourse.shopping.model.OrderCartProductsModel
