+++ app/src/test/java/woowacourse/shopping/repository/ProductRemoteRepositoryTest.kt
@@ -1,0 +1,20 @@
+package woowacourse.shopping.repository
+
+import org.junit.Before
+import org.junit.Test
+
+class ProductRemoteRepositoryTest {
+
+    @Before
+    fun setUp() {
+    }
+
+    @Test
+    fun 주문_정보를_띄울_수_있다() {
+        // given
+
+        // when
+
+        // then
+    }
+}
